fix(mdContactChips): Compile contact chips against proper scope.
After the performance modifications to make autocomplete use the
virtual repeat directive, the contact chips no longer compiled the
contents of the dropdown against the proper scope, so it was loading
a set of blank elements in the autocomplete popup.

This is a hacky way to compile the contents against the proper scope
by creating a new child scope and merging the `item` property onto it
and synching any changes (which is very important since the virtual
repeat re-uses DOM elements).

Fixes #4390.